refactor(ticket): split the Ticket god-object into facets + index ticket-number resolves
Unit 5 of the burndown. Two findings on core/models/ticket.py.

must-fix (god-object): Ticket was 1019 LOC / 47 public methods, breaching
max-public-methods=25 and silenced with a class-level PLR0904 suppression.
The cohesive instance behaviour is now composed onto five abstract-model
facets (overlay attribution, phase sessions, scheduling, evidence store,
status/introspection). The concrete Ticket keeps the fields, the FSM state
graph, the manager, and save — down to 549 LOC / 19 own public methods, and
the suppression is removed. Every method stays reachable as ticket.foo()
via the facets, so no consumer call site changes and the FSM graph is
byte-identical (pinned by test_ticket_structure).

should-fix (3e#8): _ticket_by_number loaded every non-synthetic Ticket and
filtered the derived number in Python (O(all tickets) per worktree resolve).
The forge issue number is now denormalized into an indexed issue_number
column (kept in sync by save, backfilled by migration 0004), so the resolve
is an indexed lookup + a pk-fallback for numberless tickets. The
ticket_number property is unchanged (derives from issue_url, pk fallback).

Migration 0004_ticket_issue_number adds the indexed column and backfills it;
it does not touch the squashed 0001_initial (the name-reuse gate stays green
— its current-graph frozenset gains the new name). makemigrations --check is
clean; the migration applies on a fresh DB and an old-chain DB.

Open questions & assumptions:
- Assumed: abstract-model facets (Django's native model-decomposition) are
  the right shape here rather than composed attributes. Composed attributes
  cannot preserve the flat ticket.foo() API without rewriting ~130 consumer
  call sites, which the behaviour-preserving mandate forbids; the facets keep
  every call site and the FSM intact. This is an intentional divergence from
  the "composition over mixins" style preference (the core CLAUDE.md permits
  documented divergence when following the framework's own pattern).
- Assumed: the ticket_number property must stay behaviour-identical (57
  consumers); issue_number is a query-only denormalization, not a new API.
- Decided-by-design: deferred-import pegs relocated from ticket.py to the
  facets (net -2, two type(self) conversions); the PLC0415 burndown itself
  is out of scope (Unit 20).
