+++ tests/quality/deferred_import_pegs.toml
@@ -73,7 +73,9 @@
 "src/teatree/core/models/auto_review_dispatch.py" = 3
 "src/teatree/core/models/review_loop.py" = 3
 "src/teatree/core/models/task.py" = 7
-"src/teatree/core/models/ticket.py" = 11
+"src/teatree/core/models/ticket_phase_sessions.py" = 1
+"src/teatree/core/models/ticket_scheduling.py" = 7
+"src/teatree/core/models/ticket_status.py" = 1
 "src/teatree/core/notify.py" = 4
 "src/teatree/core/on_behalf_egress.py" = 2
 "src/teatree/core/on_behalf_gate_recorded.py" = 6
